Fix many implicit 32/64-bit data conversions.
This avoids type conversion warnings in 64-bit builds.

These are mainly use of strtoul for parsing uint32_t without casting
the return value, or use of int or uint32_t when size_t is more
appropriate.

There are three APIs that we interact with that aren't using
size_t (SASL, MD5, and zlib) so we cast in and out of those.

+++ src/data-types/mailstream_low.c
@@ -93,7 +93,7 @@ static inline void mailstream_logger_internal(mailstream_low * s, int is_stream_
       f = fopen(LOG_FILE, "a"); \
       umask(old_mask); \
       if (f != NULL) { \
-        int nmemb; \
+        size_t nmemb; \
         maillock_write_lock(LOG_FILE, fileno(f)); \
         nmemb = fwrite((buf), 1, (size), f); \
         maillock_write_unlock(LOG_FILE, fileno(f)); \
@@ -120,7 +120,7 @@ static inline void mailstream_logger_internal(mailstream_low * s, int is_stream_
       f = fopen(LOG_FILE, "a"); \
       umask(old_mask); \
       if (f != NULL) { \
-        int nmemb; \
+        size_t nmemb; \
         maillock_write_lock(LOG_FILE, fileno(f)); \
         nmemb = fwrite((buf), 1, (size), f); \
         maillock_write_unlock(LOG_FILE, fileno(f)); \
